nightly.conf: add drm-tip and linux-upstream to drm_tip_repos
This way we can remove more special casing and hard-coding in dim. Until
we have updated dim, this does mean extra fetches of the repos, but that
should be a brief transitional period.

Also add git:// and https:// URLs for drm-tip, although this is likely
not going to be very useful for very many dim users.

Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>