dim: make url_to_git_url generic and rename to pick_protocol_url
No reason for the function to be so specific, make it more flexible in
case we need to choose ssh URLs later on. Also rename it
appropriately. Remove the fallback of converting ssh:// URLs to git://
URLs now that we have all the URLs with all the accepted protocols
defined in nightly.conf since drm-rerere commit 7aa4d6332234
("nightly.conf: add url lists to all repos").

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>