dim: switch to per-branch tag templates
Since we rotate maintainer functions now on a per-release basis, and
that seems to be a superior model (bettter exchanged and sharing of
knowledge) per-maintainer templates don't make much sense.

Instead use per-branch templates. To keep some separation between the
scripts and the setup I opted to put the templates into the drm-rerere
branch, where we have nightly.conf already. I've pushed a few examples
there of what I have in mind.

Once this has landed I'll send out the patch for drm-rerere with some
templates to start with.

v2: Align to the new file layout as discussed with Joonas. Comment for
function parameters.

Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Acked-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>