dim: Move to more readable fixup patch naming scheme
Because branch and repo already contain '-', move to use '__' as
the delimiter between the two for more readability.

While at it, allow explicitly specifying the branch where the
fixup should be applied in the case of a non-conflicting merge
issue.

Fixing the merge commit of drm-misc-next after noticing drm-tip
not to build would be done in the following way:

$ git diff | dim cat-to-fixup drm-misc-next

v2: Bash completion
v3: Use "fixups/" folder under rerere.

Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>