dim: Make dim sparse check each commit in the range-ish
Previously 'dim sparse' just touched all the files affected by the range
and run sparse against that. Since this may show issues from included
headers and sparse-compliance does not seem to be strongly enforced the
output is rather unhelpful.

With this change the repo is checkout at each commit, sparse is run on
incrementally, on every file that was changed.

Each result is compared to it's parent's result using remap-log to
adjust for potential line numbering changes.

Only the difference is reported.

v2:
    Clearer wording on remap-log check.
    Use the incremental build.
    Reset ref on exit.

v3:
    Make shellcheck happy:
      * Ignore SC2064 in trap
      * Quote array expansion

Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Acked-by: Daniel Vetter <daniel.vetter@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>