qf: Don't check baseline on qf pull.
At this point we are only pulling the branches so in a
distributed environment it is expected that you don't
have the latest baseline on your environment yet.

For years, all qf users were doing the initial pull
manually or with other wrap scripts. so, let's fix
it now and properly just use qf for initial pull and
start the rebase process.

Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Cc: Michel Thierry <michel.thierry@intel.com>
Cc: James Ausmus <james.ausmus@intel.com>
Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>