qf: Allow rebase function to start from anywhere.
Like we now allow qf checkout to start from an unknown
baseline we should also allow qf rebase.

But also qf rebase only starts if there are already quilt
patches applied, what doesn't make sense when you need to
start a rebase from a clean machine.

In reality what happens is that we are using checkout
or manual quilt pushes to actually rebase the patches
and the rebase function to only validate what we have
done.

By allowing the rebase to perform from anywhere we
are actually fixing the meaning of rebase and making
it more useful.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>