dim: Move maintainer-tools branch to drm-tip
Originally I wanted to move it to a completely stand-alone
maintainer-tools.git repo. But that has the problem that we need to
keep track of, and update, yet another remote. Easy way out it to
simply stuff the maintainer-tools branch into drm-tip.

The auto-upgrade hack is a bit gross, but shouldn't really cause
overhead. I've also pushed out the latest maintainer-tools (without
this patch) to drm-tip.git to make sure the patch can be tested.

The neat thing here is that this allows us to finally remove
drm-intel url hardcoding from dim.

v2: Update the docs, I've forgotten about those (Jani)

Cc: Jani Nikula <jani.nikula@linux.intel.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>