qf: fix wrong grep to remove unused patches
We would leave a trailing slash which causes patches on the git-root
directory to be removed since they don't have that slash in the series
file.

Considering we have a patches/0001-bla.patch this would lead to:

+ ...
+ for patch in patches/{*/,/}*.patch
+ grep '^/0001-bla.patch$' patches/series
+ echo No reference to patches//0001-bla.patch, deleting it.
+ rm patches//0001-bla.patch

The fix is to change the work directory to patches and stop fixing up
the path. It's also safer to just use git-ls-files so we don't
potentially remove patches still not added.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>