qf: Actually allow qf rebase from a clean environment.
The intention of commit '663e7293e042 ("qf: Allow rebase
function to start from anywhere.")' was to allow the
qf rebase to start from anywhere, including from a place
where not patch is yet applied with quilt.

However pop -f was not the right answer for that.

We need to make sure that qf pull provides a really clean
environment so qf rebase can really work from scratch to
all (-a) patches.

The goal is that a simple
qf pull -f && qf rebase drm-tip/drm-tip

Always simply work to rebase from anywhere.

v2: Implement this behavior only on -f is used to avoid losing work in
    progress (Daniel).
v3: Rebased.

Fixes: 663e7293e042 ("qf: Allow rebase function to start from anywhere.")
Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Acked-by: Lucas De Marchi <lucas.demarchi@intel.com>