qf: silence quilt refresh a little bit
Since the patches are maintained in a git branch, if we keep refreshing
we keep touching all patches with useless diff. Worse, it may change as
each developer has a different configuration. Force a unified diff
similar to git-diff in order to reduce the noise. Example:

Before (where xxxxxxxxxxx is the name of the directory):
 Index: xxxxxxxxxxx/drivers/gpu/drm/i915/intel_display.h
 ===================================================================
---- xxxxxxxxxxx.orig/drivers/gpu/drm/i915/intel_display.h       2018-03-28 11:27:14.320823971 -0700
-+++ xxxxxxxxxxx/drivers/gpu/drm/i915/intel_display.h    2018-03-28 11:27:14.312823952 -0700
+--- xxxxxxxxxxx.orig/drivers/gpu/drm/i915/intel_display.h       2018-03-28 12:11:51.916788273 -0700
++++ xxxxxxxxxxx/drivers/gpu/drm/i915/intel_display.h    2018-03-28 12:11:51.916788273 -0700

After:

	( Intentionally left blank )

With this the files that need to be updated during a refresh of the
entire pile reduces a lot.

v2: add unified helper

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>