Makefile: shut up about SC1117 and SC2207 in shellcheck
Not sure how this happened, but we gained a lot of errors. I think we should
exclude these two since we have lots of them:

SC1117: Backslash is literal in "\n". Prefer explicit escaping: "\\n".

These happen in lots of our strings (especially anything we feed to grep).

SC2207: Prefer mapfile or read -a to split command output (or quote to avoid splitting).

Our bash_completion scripts are full of these - everywhere we call compgen.

Acked-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>