dim: check all commits in dim apply-pull and push-branch
When merging a pull requests there's potentially a long list of
problematic patches. But currently we only report the first issue and
then bail out.

The upside here is that without this patch the workflow when
processing a pull is:

$ dim apply-pull ...
-> dim refuses, only reports first problem
$ dim -d apply-pull
-> you see all the issues, make up your mind to merge or not
$ dim -f apply-pull

With this patch we have one step less:

$ dim apply-pull
-> refuses pull, but prints all the issues
$ dim -f apply-pull

On the implementation:
- new helper function to check all the patches, that's the only place
  we now call warn_or_fail
- rework the check function to check for everything and return the
  final verdict

v2: Complete rework on Jani's suggestion.

v3: Wrap $@ in "" (Jani).

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Acked-by: Dave Airlie <airlied@redhat.com>
Cc: Jani Nikula <jani.nikula@linux.intel.com>
Cc: Dave Airlie <airlied@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>