dim: auto-commit in apply-pull if rerere solved it all
Now that we experiment with dim for drm-next it's much more likely
that pull requests have conflicts. But also that dim already knows
about them, in the recent drm-intel-next pull it resolve 7/8
conflicts.

If it solves them all then just go ahead an commit.

v2: Conflict howto is now in drm-tip.rst (Jani).

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Cc: Dave Airlie <airlied@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>