dim: check for empty pulls
This could be a parsing error of the pull (for all the people who
don't use a script like dim pull-request to make sure all the silly
details are right), or some confusion on the part of the sender, or
something else.

Either way not good to continue.

v2: Use -z and drop the wc -l (Jani).

Cc: Dave Airlie <airlied@gmail.com>
Acked-by: Jani Nikula <jani.nikula@linux.intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>