dim: remove all mentions of drm-upstream
The one case in dim_status is entirely unused.

The usage in dim_setup seems still in use, but this is code for
backwards compat hacks predating the original creating of the
maintainer-tools branch - i.e. no maintainer or committer except me
ever needed this.

There's also no need to make sure drm-tip has all the right remotes
nowadays, because that's all encoded in nightly.conf, and will be
checked (if the remotes are missing) in dim_rebuild_tip.

Reported-by: Daniel Stone <daniels@collabora.com>
Cc: Daniel Stone <daniels@collabora.com>
Acked-by: Daniel Stone <daniels@collabora.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>