doc: clarify the target audience for the status command
'dim status' is intended to be used by maintainers and not by commiters.
Move the description of the subcommand into the 'COMMANDS FOR MAINTAINERS'
area and clarify in the COMMITERS area that they should use vanilla
'git status' for checking the health of the branch.

v2: Fix duplicate 'status' warnings by merging the text into the 'COMMANDS
FOR MAINTAINERS' description.

Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>