dim: allow tags in backmerges
Linus prefers that we backmerge a specific tag instead of a random
point in his branch. Allow that.

I guess it'd be nice to somehow figure out where a tag came from, but
git doesn't namespace tags. So that idea is out of the window
unfortunately.

v2: Unlazy and also update the docs.

v3: Make the check work from anywhere, we need to move the cd a bit
up (Dave).

Requested-by: Dave Airlie <airlied@redhat.com>
Cc: Dave Airlie <airlied@redhat.com>
Tested-by: Dave Airlie <airlied@redhat.com>
Reviewed-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>