nightly.conf: drop amd branches from drm-tip
We've had a stale repo for amd in drm-tip since around v4.15 i.e. for
more than three years. Nobody seems to notice or care. Drop the amd
branches from drm-tip.

Having the current amd branches in drm-tip would be nice to have, if
only to have a common drm integration tree. However, maintaining that
has a cost due to the inevitable conflicts. We can add the branches back
if and when there's interest in sharing the burden.

Cc: Alex Deucher <alexander.deucher@amd.com>
Cc: Christian König <christian.koenig@amd.com>
Cc: Pan, Xinhui <Xinhui.Pan@amd.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Acked-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Reviewed-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>